Revert "Merge pull request #19 from abhinathtiwari/dev"
This reverts commit b4c3c2963835c8dbf299f1959086d0080c568bc0, reversing
changes made to 0ec1a33d3fd007e78dd84014f34de517d88a769e.
