+++ backend/src/config/youtubeAuth.ts
@@ -1,4 +1,3 @@
-/** YouTube OAuth2 client configuration and shared instance. */
 import { google } from "googleapis";
 
 export const oauth2Client = new google.auth.OAuth2(
