+++ backend/src/jobs/reminder.job.ts
@@ -1,4 +1,3 @@
-/** Weekly cron job to send reminders and AI ideas to inactive creators. */
 import cron from "node-cron";
 import { User } from "../models/Users";
 import { google } from "googleapis";
