+++ backend/src/routes/authUser.routes.ts
@@ -1,4 +1,3 @@
-/** Express routes handling user authentication and registration logic. */
 import { Router } from "express";
 import { User } from "../models/Users";
 import bcrypt from "bcrypt";
