+++ backend/src/routes/youtube.cron.routes.ts
@@ -1,4 +1,3 @@
-/** Express routes for managing and triggering the SEO cron job. */
 import { Router } from "express";
 import { google } from "googleapis";
 import { oauth2Client } from "../config/youtubeAuth";
