+++ backend/src/server.ts
@@ -1,4 +1,3 @@
-/** Entry point of the server; initializes database connections and starts listening. */
 import dotenv from "dotenv";
 dotenv.config();
