--- backend/src/utils/errorHandler.ts
@@ -1,51 +1,0 @@
-/** Global error handler middleware and custom error class for standardized responses. */
-import { Request, Response, NextFunction } from "express";
-
-export class AppError extends Error {
-    public statusCode: number;
-    public status: string;
-    public isOperational: boolean;
-
-    constructor(message: string, statusCode: number) {
-        super(message);
-        this.statusCode = statusCode;
-        this.status = `${statusCode}`.startsWith("4") ? "fail" : "error";
-        this.isOperational = true;
-
-        Error.captureStackTrace(this, this.constructor);
-    }
-}
-
-export const globalErrorHandler = (
-    err: any,
-    req: Request,
-    res: Response,
-    next: NextFunction
-) => {
-    err.statusCode = err.statusCode || 500;
-    err.status = err.status || "error";
-
-    if (process.env.NODE_ENV === "development") {
-        res.status(err.statusCode).json({
-            status: err.status,
-            error: err,
-            message: err.message,
-            stack: err.stack,
-        });
-    } else {
-        // Production mode
-        if (err.isOperational) {
-            res.status(err.statusCode).json({
-                status: err.status,
-                message: err.message,
-            });
-        } else {
-            // Programming or other unknown error: don't leak error details
-            console.error("ERROR 💥", err);
-            res.status(500).json({
-                status: "error",
-                message: "Something went very wrong!",
-            });
-        }
-    }
-};
