+++ backend/src/utils/growthAnalysis.ts
@@ -1,4 +1,3 @@
-/** Utility for tracking upload consistency and generating growth-oriented AI analysis. */
 import { google } from "googleapis";
 import axios from "axios";
 import { User } from "../models/Users";
