+++ backend/src/utils/middlewares.ts
@@ -1,4 +1,3 @@
-/** Custom Express middlewares for user authentication and authorization. */
 import { Request, Response, NextFunction } from "express";
 import jwt from "jsonwebtoken";
 import { User } from "../models/Users";
