+++ backend/src/utils/prompts.ts
@@ -1,4 +1,3 @@
-/** Factory for building structured AI prompt templates for video and growth analysis. */
 
 export function buildGrowthPrompt(videos: any[]) {
   return `
