+++ frontend/src/App.tsx
@@ -1,4 +1,3 @@
-/** Root React component managing global routing and application structure. */
 import { BrowserRouter, Routes, Route, Navigate } from "react-router-dom";
 import Auth from "./pages/Auth";
 import Home from "./pages/Home";
