+++ frontend/src/api.ts
@@ -1,4 +1,3 @@
-/** Centralized Axios instance with base configuration for all API requests. */
 import axios from "axios";
 
 const api = axios.create({
