+++ frontend/src/pages/User.tsx
@@ -1,4 +1,3 @@
-/** User profile management page (Placeholder/Development). */
 import { useEffect, useState } from "react";
 import Layout from "../components/Layout";
 import api from "../api";
